backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.4b2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
Deprecated==1.2.12
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-nose==1.4.7
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.1.2
gitdb==4.0.7
GitPython==3.1.17
idna==2.10
mypy-extensions==0.4.3
nose==1.3.7
numpy==1.20.2
pandas==1.2.4
pathspec==0.8.1
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
six==1.15.0
smmap==4.0.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
